Toevoegen van ProgramFlow classs voor opslaan van programmalogica.
